chowa-stylebot/requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ── Validation ─────────────────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1

# ── HTTP client (weather API, Vision API) ──────────────────────────────────────
httpx==0.27.0

# ── Image processing ───────────────────────────────────────────────────────────
rembg==2.0.56
Pillow==10.3.0

# ── Cloud storage ──────────────────────────────────────────────────────────────
boto3==1.34.84                   # AWS S3
google-cloud-storage==2.16.0    # GCS (optional, pick one)
google-cloud-vision==3.7.2      # Vision API for garment tagging

# ── Auth ───────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# ── Config ─────────────────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Testing ────────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx                            # also used in tests as async client
